fix(refactoring): added routes controller
