+++ controller/index.ts
@@ -1,0 +1,5 @@
+export * as commonController from './Common';
+export * as messageController from './Messages';
+export * as adminController from './Admin';
+export * as customerController from './Customer';
+export * as serviceManController from './Serviceman';
